restructuring module for packaging
