+++ pyproject.toml
@@ -1,0 +1,6 @@
+[build-system]
+requires=[
+  "setuptools>=42",
+  "wheel"
+  ]
+build-backend="setuptools.build_meta"
