+++ setup.cfg
@@ -1,0 +1,21 @@
+[metadata]
+name = spilite
+version = 0.1
+author = xdylanm
+description = A lite wrapper for spidev with additional chip select options
+long_description = file:README.md
+long_description_content = text/markdown
+url = https://github.com/xdylanm/spilite
+classifiers = 
+  Programming Language :: Python :: 3
+  License :: OSI Approved :: MIT License
+  Operating System :: OS Independent
+
+[options]
+package_dir = 
+  = src
+packages = find:
+python_requires = >=3.6
+
+[options.packages.find]
+where = src
